module_camelot_for_pdf_tables/a00_intro_camelot/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Project description"
authors = ["Bhishan Poudel"]

[tool.poetry.dependencies]
python = "^3.7"
click = ">=6.7"
matplotlib = "2.2.3"
numpy = "1.13.3"
opencv-python = "3.4.2.17"
openpyxl = "2.5.8"
pandas = "0.23.4"
PyPDF2 = "1.26.0"
camelot-py = { version = "*", extras = ["cv"] }


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"